% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AlphaWords}
\alias{AlphaWords}
\title{English word dictionary}
\format{
\subsection{\code{AlphaWords}}{

A data frame with 370.104 rows and 1 column:
\describe{
\item{words}{word}
}
}
}
\source{
\url{https://github.com/dwyl/english-words/blob/master/words.txt}
}
\usage{
AlphaWords
}
\description{
A large list of English words
}
\keyword{datasets}
