% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CommonWords}
\alias{CommonWords}
\title{Common English word dictionary}
\format{
\subsection{\code{CommonWords}}{

A data frame with 4.318 rows and 1 column:
\describe{
\item{words}{word}
}
}
}
\source{
<>
}
\usage{
CommonWords
}
\description{
A small list of common English words
}
\keyword{datasets}
