% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_words.R
\name{join_words}
\alias{join_words}
\title{Probability of word}
\usage{
join_words(word, dictionary_hash, total, max_word_length)
}
\arguments{
\item{word}{Any given single fussed/misspelt word.}

\item{dictionary_hash}{Hashed dictionary obtained through the train_viterbi() function.}

\item{total}{Total sum of dictionary values.}

\item{max_word_length}{Longest word in the dictionary.}
}
\value{
Returns the corrected word/words from the provided word.
}
\description{
Checks how common the word is in the dictionary provided.
}
