% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIS4370.Final.Project-package.R
\docType{package}
\name{LIS4370.Final.Project-package}
\alias{LIS4370.Final.Project}
\alias{LIS4370.Final.Project-package}
\title{LIS4370.Final.Project: Viterbi NLP Preprocessing}
\description{
This package is currently a viterbi preprocessing tool that is intended to primarly clean data for NLP word. The tool corrects misspelt and fussed words in corrupted text by training a dictionary and comparing word probability to said dictionary. The algorithm is not perfect but can be quite powerful when used in the right circumstances.
}
\author{
\strong{Maintainer}: Lilith Holland \email{Lilith.Holland@example.com}

}
\keyword{internal}
