% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_dictionary.R
\name{populate_dictionary}
\alias{populate_dictionary}
\title{Dictionary initialization}
\usage{
populate_dictionary(dictionary, remove_singles = FALSE)
}
\arguments{
\item{dictionary}{Any char data structure that can be run through sapply().}

\item{remove_singles}{Bool flag to remove single letter words from dictionary.}
}
\value{
Processed dictionary.
}
\description{
Prepares a dictionary for use with the viterbi algorithm.
}
