% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleText}
\alias{SampleText}
\title{Sample sentence for viterbi segment}
\format{
\subsection{\code{SampleText}}{

Character vector of a few sample sentences:
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
SampleText
}
\description{
A few sentences with errors to demonstrate the viterbi segment
}
\keyword{datasets}
