% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_viterbi.R
\name{train_viterbi}
\alias{train_viterbi}
\title{Train viterbi dictionary}
\usage{
train_viterbi(text, dictionary, hashed_dictionary = NULL)
}
\arguments{
\item{text}{Any sentence/paragraph/paper or series of words separated by a space or punctuation.}

\item{dictionary}{Any char data structure that has been processed by dictionary()}

\item{hashed_dictionary}{Hashed dictionary obtained through the train_viterbi() function.}
}
\value{
Hashed_dictionary, total, and max_word_length as a list
}
\description{
Builds and trains the hashed dictionary for use with the viterbi algorithm.
}
