% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_segment.R
\name{viterbi_segment}
\alias{viterbi_segment}
\title{Viterbi segment algorithm}
\usage{
viterbi_segment(text, dictionary_hash, total, max_word_length)
}
\arguments{
\item{text}{Any given single fussed text.}

\item{dictionary_hash}{Hashed dictionary obtained through the train_viterbi() function.}

\item{total}{Total sum of dictionary values.}

\item{max_word_length}{Longest word in the dictionary.}
}
\value{
Returns he decomposed word and probability of correctness as a list..
}
\description{
Runs the viterbi segment algorithm on any text. The algorithm will use the provided dictionary to split and correct fussed and misspelt words.
}
