% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_prob.R
\name{word_prob}
\alias{word_prob}
\title{Word probability}
\usage{
word_prob(word, dictionary_hash, total)
}
\arguments{
\item{word}{Any given single word.}

\item{dictionary_hash}{Hashed dictionary obtained through the train_viterbi() function.}

\item{total}{Total sum of dictionary values.}
}
\value{
Returns simple probability of word showing up in trained dictionary.
}
\description{
Checks the probability of a word showing up based on a trained dictionary.
}
