% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\name{words}
\alias{words}
\title{Word split}
\usage{
words(text)
}
\arguments{
\item{text}{Any sentence/paragraph/paper or series of words separated by a space or punctuation.}
}
\value{
Returns the separated text in the form of individual words.
}
\description{
Separates a sentence into individual words removing numbers.
}
